Massive overhaul on role resources

+++ backend/src/main/java/com/caiohbs/crowdcontrol/model/Role.java
@@ -63,7 +63,7 @@ public class Role {
         return salary;
     }
 
-    public void setSalary(int salary) {
+    public void setSalary(double salary) {
         this.salary = salary;
     }
